Finish modularization
