+++ app/src/androidTest/java/com/maximo/douglas/githubconsumer/testutils/FileUtils.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.utils
+package com.maximo.douglas.githubconsumer.testutils
 
 object FileUtils {
