+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
-        android:name="com.maximo.douglas.githubconsumer.GithubApiConsumerApp"
+        android:name="com.maximo.douglas.githubconsumer.Application"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -14,7 +14,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.maximo.douglas.githubconsumer.MainActivity">
+        <activity android:name="com.maximo.douglas.githubconsumer.ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
