+++ app/src/main/java/com/maximo/douglas/githubconsumer/Application.kt
@@ -1,13 +1,13 @@
 package com.maximo.douglas.githubconsumer
 
 import android.app.Application
-import com.maximo.douglas.githubconsumer.di.othersModule
+import com.maximo.douglas.commons.di.commonsModule
 import com.maximo.douglas.githubconsumer.di.remoteDataSourceModule
 import com.maximo.douglas.githubconsumer.di.serviceModule
 import com.maximo.douglas.githubconsumer.di.viewModelModule
 import org.koin.core.context.startKoin
 
-class GithubApiConsumerApp : Application() {
+class Application : Application() {
 
     override fun onCreate() {
         startKoin {
@@ -15,7 +15,7 @@ class GithubApiConsumerApp : Application() {
                 listOf(
                     viewModelModule,
                     remoteDataSourceModule,
-                    othersModule,
+                    commonsModule,
                     serviceModule
                 )
             )
