+++ app/src/main/res/layout/main_activity.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_activity_constraint_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.maximo.douglas.githubconsumer.MainActivity">
+    tools:context="com.maximo.douglas.githubconsumer.ui.MainActivity">
 
     <fragment
         android:id="@+id/navHostFragment"
