+++ app/src/test/java/com/maximo/douglas/githubconsumer/components/gitpullrequestcard/context/GitPullRequestCardTest.kt
@@ -5,7 +5,7 @@ import android.view.View
 import com.maximo.douglas.commons.utils.StringUtils.compactStringWithDots
 import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
 import com.maximo.douglas.domain.entities.user.GitUser
-import com.maximo.douglas.githubconsumer.components.utils.BaseComponentTest
+import com.maximo.douglas.githubconsumer.testutils.BaseComponentTest
 import com.maximo.douglas.githubconsumer.ui.components.pullrequestcard.GitPullRequestCardComponent
 import kotlinx.android.synthetic.main.git_pull_request_card_component.view.*
 import org.junit.Assert.assertEquals
