+++ app/src/test/java/com/maximo/douglas/githubconsumer/components/gitrepositorycard/context/GitRepositoryCardTest.kt
@@ -5,7 +5,7 @@ import android.view.View
 import com.maximo.douglas.commons.utils.StringUtils.compactStringWithDots
 import com.maximo.douglas.domain.entities.gitrepository.GitRepository
 import com.maximo.douglas.domain.entities.user.GitUser
-import com.maximo.douglas.githubconsumer.components.utils.BaseComponentTest
+import com.maximo.douglas.githubconsumer.testutils.BaseComponentTest
 import com.maximo.douglas.githubconsumer.ui.components.repositorycard.GitRepositoryCardComponent
 import kotlinx.android.synthetic.main.git_repository_card_component.view.*
 import org.junit.Assert.assertEquals
