+++ app/src/test/java/com/maximo/douglas/githubconsumer/testutils/TestContextProvider.kt
@@ -1,9 +1,10 @@
-package com.maximo.douglas.githubconsumer.viewmodels.utils
+package com.maximo.douglas.githubconsumer.testutils
 
+import com.maximo.douglas.commons.utils.CoroutineContextProvider
 import kotlinx.coroutines.Dispatchers
 import kotlin.coroutines.CoroutineContext
 
-class TestContextProvider : com.maximo.douglas.commons.utils.CoroutineContextProvider() {
+class TestContextProvider : CoroutineContextProvider() {
 
     override val Main: CoroutineContext = Dispatchers.Unconfined
     override val IO: CoroutineContext = Dispatchers.Unconfined
