+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/gitpullrequest/context/SelectGitPullRequestViewModelTest.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.Observer
 import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel.SelectGitPullRequestViewModelState
-import com.maximo.douglas.githubconsumer.viewmodels.utils.TestContextProvider
-import com.maximo.douglas.githubconsumer.viewmodels.utils.TestCoroutineRule
-import com.maximo.douglas.githubconsumer.viewmodels.utils.faker.GitPullRequestFaker
+import com.maximo.douglas.githubconsumer.testutils.TestContextProvider
+import com.maximo.douglas.githubconsumer.testutils.TestCoroutineRule
+import com.maximo.douglas.githubconsumer.testutils.faker.GitPullRequestFaker
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Before
 import org.junit.Rule
