+++ commons/build.gradle
@@ -73,4 +73,8 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.11.0'
     kapt 'com.github.bumptech.glide:compiler:4.12.0'
     implementation 'jp.wasabeef:glide-transformations:4.1.0'
+
+    // KOIN
+    implementation "io.insert-koin:koin-android:3.0.1"
+    testImplementation "io.insert-koin:koin-test-junit4:3.0.1"
 }
